backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Configuration
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Excel file processing
openpyxl==3.1.2
xlrd==2.0.1

# Document processing
python-docx==1.1.0

# File upload handling
python-multipart==0.0.6

# LLM Integration
anthropic==0.40.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
jellyfish==1.0.3  # For fuzzy string matching (soundex, etc.)
